+++ packages/components/src/components/ui/Account/CurrentAccountsHeader/CurrentAccountsHeader .tsx
@@ -1,7 +1,4 @@
-import { AnimatePresence, LayoutGroup } from 'framer-motion'
-
 import { Button, Icon } from '../../atoms'
-import LayoutWrapper from '../../utils/LayoutWrapper'
 
 export interface CurrentAccountsHeaderProps {
   onCreatePaymentAccount: () => void
@@ -13,16 +10,10 @@ const CurrentAccountsHeader = ({ onCreatePaymentAccount }: CurrentAccountsHeader
       <span className="md:pl-4 leading-8 font-medium text-2xl md:text-[1.75rem]">
         Currents accounts
       </span>
-      <LayoutGroup>
-        <AnimatePresence initial={false}>
-          <LayoutWrapper>
-            <Button size="big" onClick={onCreatePaymentAccount} variant="secondary">
-              <Icon name="add/16" className="text-default-text" />
-              <span className="pl-2 md:inline-block">New Account</span>
-            </Button>
-          </LayoutWrapper>
-        </AnimatePresence>
-      </LayoutGroup>
+      <Button size="big" onClick={onCreatePaymentAccount} variant="secondary" className="w-[172px]">
+        <Icon name="add/16" className="text-default-text" />
+        <span className="pl-2 md:inline-block">New Account</span>
+      </Button>
     </div>
   )
 }
